feat(layer): final refacto
