+++ antarest/study/business/layer_management.py
@@ -11,15 +11,10 @@
 # This file is part of the Antares project.
 from typing import List
 
-from antarest.core.exceptions import LayerNotFound
-from antarest.study.business.area_management import logger
-from antarest.study.business.areas.area_utils import _get_area_layers
 from antarest.study.business.model.layer_model import Layer, LayerCreation, LayerUpdate
 from antarest.study.business.study_interface import StudyInterface
 from antarest.study.storage.variantstudy.model.command.create_layer import CreateLayer
-from antarest.study.storage.variantstudy.model.command.icommand import ICommand
 from antarest.study.storage.variantstudy.model.command.remove_layer import RemoveLayer
-from antarest.study.storage.variantstudy.model.command.update_config import UpdateConfig
 from antarest.study.storage.variantstudy.model.command.update_layer_name import UpdateLayerName
 from antarest.study.storage.variantstudy.model.command_context import CommandContext
 
@@ -47,70 +42,6 @@ class LayerManager:
             study_version=study.version,
         )
         study.add_commands([command])
-
-    def update_layer_areas(self, study: StudyInterface, layer_id: str, areas: List[str]) -> None:
-        logger.info(f"Updating layer {layer_id} with areas {areas}")
-        file_study = study.get_files()
-        layers = file_study.tree.get(["layers", "layers", "layers"])
-        if layer_id not in [str(layer) for layer in list(layers.keys())]:
-            raise LayerNotFound
-        areas_ui = file_study.tree.get(["input", "areas", ",".join(file_study.config.areas), "ui"])
-        # standardizes 'areas_ui' to a dictionary format even if only one area exists.
-        cfg_areas = list(file_study.config.areas)
-        if len(cfg_areas) == 1:
-            areas_ui = {cfg_areas[0]: areas_ui}
-
-        existing_areas = [
-            area for area in areas_ui if "ui" in areas_ui[area] and layer_id in _get_area_layers(areas_ui, area)
-        ]
-        to_remove_areas = [area for area in existing_areas if area not in areas]
-        to_add_areas = [area for area in areas if area not in existing_areas]
-        commands: List[ICommand] = []
-
-        def create_update_commands(area_id: str) -> List[ICommand]:
-            return [
-                UpdateConfig(
-                    target=f"input/areas/{area_id}/ui/layerX",
-                    data=areas_ui[area_id]["layerX"],
-                    command_context=self._command_context,
-                    study_version=study.version,
-                ),
-                UpdateConfig(
-                    target=f"input/areas/{area_id}/ui/layerY",
-                    data=areas_ui[area_id]["layerY"],
-                    command_context=self._command_context,
-                    study_version=study.version,
-                ),
-                UpdateConfig(
-                    target=f"input/areas/{area_id}/ui/ui/layers",
-                    data=areas_ui[area_id]["ui"]["layers"],
-                    command_context=self._command_context,
-                    study_version=study.version,
-                ),
-            ]
-
-        for area in to_remove_areas:
-            area_to_remove_layers: List[str] = _get_area_layers(areas_ui, area)
-            if layer_id in areas_ui[area]["layerX"]:
-                del areas_ui[area]["layerX"][layer_id]
-            if layer_id in areas_ui[area]["layerY"]:
-                del areas_ui[area]["layerY"][layer_id]
-            if layer_id in area_to_remove_layers:
-                areas_ui[area]["ui"]["layers"] = " ".join(
-                    [area_layer for area_layer in area_to_remove_layers if area_layer != layer_id]
-                )
-            commands.extend(create_update_commands(area))
-        for area in to_add_areas:
-            area_to_add_layers: List[str] = _get_area_layers(areas_ui, area)
-            if layer_id not in areas_ui[area]["layerX"]:
-                areas_ui[area]["layerX"][layer_id] = areas_ui[area]["ui"]["x"]
-            if layer_id not in areas_ui[area]["layerY"]:
-                areas_ui[area]["layerY"][layer_id] = areas_ui[area]["ui"]["y"]
-            if layer_id not in area_to_add_layers:
-                areas_ui[area]["ui"]["layers"] = " ".join(area_to_add_layers + [layer_id])
-            commands.extend(create_update_commands(area))
-
-        study.add_commands(commands)
 
     def remove_layer(self, study: StudyInterface, layer_id: str) -> None:
         """
